# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.KnowledgeOrgSelectReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
    required:
    - knowledgeId
    type: object
...
  request.KnowledgeTransferUserAdminReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeUser:
        allOf:
        - $ref: '#/definitions/request.KnowledgeUserInfo'
        description: 知识库用户信息
      permissionId:
        description: 权限id编辑时传入
        type: string
    required:
    - knowledgeId
    - knowledgeUser
    type: object
  request.KnowledgeUserAddReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeUserList:
        description: 知识库用户信息
        items:
          $ref: '#/definitions/request.KnowledgeUserInfo'
        type: array
      permissionType:
        description: '权限类型:0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑 授权权限>编辑权限>查看权限'
        type: integer
    required:
    - knowledgeId
    - knowledgeUserList
    type: object
  request.KnowledgeUserDeleteReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      permissionId:
        description: 知识库用户信息权限信息
        type: string
    required:
    - knowledgeId
    - permissionId
    type: object
  request.KnowledgeUserEditReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      knowledgeUserList:
        description: 知识库用户信息
        items:
          $ref: '#/definitions/request.KnowledgeUserInfo'
        type: array
    required:
    - knowledgeId
    - knowledgeUserList
    type: object
  request.KnowledgeUserInfo:
    properties:
      orgId:
        type: string
      permissionId:
        description: 权限id编辑时传入
        type: string
      permissionType:
        description: '权限类型: -1 删除此用户权限；0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑
          授权权限>编辑权限>查看权限'
        type: integer
      userId:
        type: string
    required:
    - orgId
    - userId
    type: object
  request.KnowledgeUserNoPermitSelectReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
      orgId:
        description: 选择组织id
        type: string
      transfer:
        description: 是否是转让列表
        type: boolean
    required:
    - knowledgeId
    - orgId
    type: object
  request.KnowledgeUserSelectReq:
    properties:
      knowledgeId:
        description: 知识库id
        type: string
    required:
    - knowledgeId
    type: object
...
  response.KnowOrgInfo:
    properties:
      orgId:
        type: string
      orgName:
        type: string
    type: object
  response.KnowOrgUserInfoResp:
    properties:
      orgId:
        type: string
      orgName:
        type: string
      userInfoList:
        items:
          $ref: '#/definitions/response.KnowUserInfo'
        type: array
    type: object
  response.KnowUserInfo:
    properties:
      userId:
        type: string
      userName:
        type: string
    type: object
...
  response.KnowledgeUserInfo:
    properties:
      orgId:
        type: string
      orgName:
        type: string
      permissionId:
        type: string
      permissionType:
        description: '权限类型: -1 删除此用户权限；0: 查看权限; 10: 编辑权限; 20: 授权权限,数值不连续的原因防止后续有中间权限，目前逻辑
          授权权限>编辑权限>查看权限'
        type: integer
      userId:
        type: string
      userName:
        type: string
    type: object
  response.KnowledgeUserPermissionResp:
    properties:
      knowledgeUserInfoList:
        items:
          $ref: '#/definitions/response.KnowledgeUserInfo'
        type: array
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- |
    --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n|
    X-Org-Id      | 组织ID    |\r\n| X-Client-Id   | 客户端标识|\r\n\r\n## HTTP Status\r\n|
    HTTP Status             | 说明                   |\r\n| -----------------------
    | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n|
    400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败
    \           |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n|
    一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n|
    guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      |
    【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n|
    permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      |
    权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n##
    `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n
    \   \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\":
    \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n
    \       {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n
...
    \"操作成功\"\r\n}\r\n```"
...
  version: v0.0.1
paths:
...
  /knowledge/org:
    get:
      consumes:
      - application/json
      description: 知识库组织列表
      parameters:
      - description: 知识库组织列表请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeOrgSelectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowOrgInfo'
              type: object
      security:
      - JWT: []
      summary: 知识库组织列表
      tags:
      - knowledge
...
  /knowledge/user:
    get:
      consumes:
      - application/json
      description: 知识库用户权限列表
      parameters:
      - description: 知识库用户权限列表请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserSelectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowledgeUserPermissionResp'
              type: object
      security:
      - JWT: []
      summary: 知识库用户权限列表
      tags:
      - knowledge
  /knowledge/user/add:
    post:
      consumes:
      - application/json
      description: 增加知识库用户
      parameters:
      - description: 增加知识库用户请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 增加知识库用户
      tags:
      - knowledge
  /knowledge/user/admin/transfer:
    post:
      consumes:
      - application/json
      description: 转让管理员权限
      parameters:
      - description: 转让管理员权限请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeTransferUserAdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 转让管理员权限
      tags:
      - knowledge
  /knowledge/user/delete:
    delete:
      consumes:
      - application/json
      description: 删除知识库用户
      parameters:
      - description: 删除知识库用户请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 删除知识库用户
      tags:
      - knowledge
  /knowledge/user/edit:
    post:
      consumes:
      - application/json
      description: 修改知识库用户
      parameters:
      - description: 修改知识库用户请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: 修改知识库用户
      tags:
      - knowledge
  /knowledge/user/no/permit:
    get:
      consumes:
      - application/json
      description: 没有知识库用户列表
      parameters:
      - description: 没有知识库用户列表请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.KnowledgeUserNoPermitSelectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.KnowOrgUserInfoResp'
              type: object
      security:
      - JWT: []
      summary: 没有知识库用户列表
      tags:
      - knowledge
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"